[Stdlib] Be a bit more deffensive when computing _word_index
It could be that the UInt does not fit into the Int, but since we are
shifting it right by a constant, we know that it will fit. So switch the
casts to be more defensive.

MODULAR_ORIG_COMMIT_REV_ID: d0c62a0cd3e77656f0abd510d77d9f4d16db86e7