[SDK] Refactor MLIR value remapping in Graph class (#61168)
This change should be a NFC, but it manages to also prevent a hang in the `benchmark_pipeline_latency.py` script for some reason.

The overall change is an improvement in code readability, so just moving on for now.

Fixes PERF-923

MAX_GRAPH_API_ORIG_REV_ID: 9ae536f81d1b971e39fe470dfef805779f9cf306