[Graph API] Remove _reconcile_weights
This was a utility that was used to fix up some tests when we were in a
transition period between implicit and explicit devices.

It inserts transfers between driver weights to match what is expected by
the graph.

It is no longer needed (as all allocations in the weight dict should be
on correct device) so remove it.

MAX_GRAPH_API_ORIG_REV_ID: 6a930b627b8ec2fb72b73b4329a7ef3b0a8daa13