[stdlib] Remove `regex` module
This was an experiment that wasn't meant for the `stdlib`. To avoid any
confusion, just remove it for now. We are not actively working on this
or investing in this area, so this should help eliminate any concerns
around API design, evolution, Python naming, performance, correctness,
etc.

MODULAR_ORIG_COMMIT_REV_ID: da6545e674033078f444b236387ece1817f4398f