[External] [stdlib] Fix `atof` precision issues and use SOTA algorithm (#61913)
[External] [stdlib] Fix `atof` precision issues and use SOTA algorithm

## Goal

Fix the correctness problem of the `atof` function.

Fixes the following issue:
* https://github.com/modularml/mojo/issues/3419

Supersedes the following PR:
* https://github.com/modularml/mojo/pull/3422

This pull request is a complete reimplementation of the `atof` function,
which was returning incorrect numbers due to doing too many floating
point operations and loosing accuracy in the process.

The primary goal of this PR was to provide a correct implementation and
raising an error when the function cannot handle the input. Since this
require a full rewriting and a quite complexe parsing, let's use a state
of the art algorithm while we're at it.

The rational for going with state of the art directly is also that this
function will be used massively everywhere when parsing json, xml, yaml,
user inputs, csv... Thus performance is very important here. While we
can't provide a very fast `atof` right now, we want to lay out the
groundwork to make this possible.

As a side note, the error messages were improved.

## Implementation

The algorithm chosen here was the one from the following papers:
* [Number Parsing at a Gigabyte per
Second](https://arxiv.org/abs/2101.11408) by Daniel Lemire
* [Fast Number Parsing Without
Fallback](https://arxiv.org/abs/2212.06644) by Noble Mushtak and Daniel
Lemire

The reference implementation used for this PR can be found here:
https://github.com/CarlVerret/csFastFloat

The part about parsing digits was not taken from the paper and was
written by myself, with a focus on using SIMD and avoid hard to predict
branches.
## Limitations

### Performance
I'm not sure it would be interesting to talk about the performance
benefits of this pull request because:
1) The original `atof` was returning wrong values, and it's no use
running a benchmark against an incorrect function.
2) We have a few issues in the stdlib which tank the performance of this
function, notably `String.removesuffix()` which returns a `String`
instead of a `StringSlice` and the fact that `String` doesn't have short
string optimization yet, meaning that calling `String.removesuffix()` do
a memory allocation on the heap.
3) To avoid many instructions, we need the intrinsics to do a
multiplication (`full_multiplication` function) which takes two 64 bits
integers and returns two 64 bits integers (high bits and low bits).
4) Many constants were hardcoded in this pull request and could be
adapted to different architectures and benchmarked independently. I'm
notably thinking about the function converting a serie of digit
characters into an `UInt64`. Generating those constants at compile-time
is not possible yet as Mojo's compile-time interpreter has some progress
to do first.

I expect that when more progress have been done on those issues, we can
start benchmarking the `atof` function seriously and look into further
optimization. The base algorithm is there, as well as the unit tests and
so it should make the optimization step very easy.

### Correctness
We can't currently process floating point representation which have a
decimal significand that doesn't fit in an `UInt64`. e.g.
`47421763.548648646474532187448684` because
`47421763548648646474532187448684` doesn't fit in a `UInt64`.

Rather than returning a wrong number, we produce an error.

This limitation can surely be removed in the future, but was deemed
out-of-scope for this pull request.

### Merging this PR

* While the diff can seems big, many of the lines added are constants,
comments and licenses.
* We can have an open discussion with the maintainers about how to split
this PR. As you might know, I very much like small PRs too, but this is
also good to have a bird view of the situation with this global PR. If
you want to split it and have a specific split in mind, feel free to
tell me.

ORIGINAL_AUTHOR=Gabriel de Marmiesse gabrieldemarmiesse@gmail.com
PUBLIC_PR_LINK=modular/modular#3434

Co-authored-by: Gabriel de Marmiesse <gabrieldemarmiesse@gmail.com>
Closes modular/modular#3434
MODULAR_ORIG_COMMIT_REV_ID: ffa2547b235ee7487e4ef7a94ca09eb2751b3105