[Kernels] Add `update_occurrences_tensor` kernel
This kernel is used to update the occurrences tensor when a new token is added to the prompt.
The occurrences tensor is a 2D tensor where the first column contains the token ids and the second column contains the counts of how many times each token has appeared in the prompt.

MODULAR_ORIG_COMMIT_REV_ID: 054c70ffdfc96e35b6e9466e53078fcb53910767