[External] [stdlib] Fix block-level sum and prefix sum operations (#61649)
[External] [stdlib] Fix block-level sum and prefix sum operations

Block-level `sum` and `prefix_sum` are not currently implemented
correctly. The problems are:

- Multiple places where the execution is blocked (`warp_id()` and
`lane_id()` broadcast the values to all warp lanes, but only one of the
lanes enters the block where these functions are called - as the result,
synchronize operations are blocking the execution because every thread
other than the leader is not joining the execution)
- Incorrect `exclusive` handling

This commit is a part of the project submission for "Modular GPU Kernel
Hackathon" at the AGI House.

Co-authored-by: Kirill Bobyrev <3352968+kirillbobyrev@users.noreply.github.com>
Closes modular/modular#4553
MODULAR_ORIG_COMMIT_REV_ID: b44e6f48f1f89451efb2785faecd5f8548f7d524