[stdlib] Remove `os.abort()` for `Writable` messages
By having support for `os.abort()` with a variadic number of `Writable`
messages, we force every caller to elaborate a bunch of IR for `String`
along with the writer abstractions like `write_buffered`,
`_WriteBufferStack`, and friends.

Instead, narrow the API of `os.abort()` to just take a single `String`
argument for the message. This pushes the complexity and responsibility
to client code to create the `String` (perhaps using the writer or
format interfaces if desirable) before calling `os.abort(message)`. In
general, this provides a nice simple API and minimizes the amount of IR
generated which is a win for compile times.

MODULAR_ORIG_COMMIT_REV_ID: 5636b6404724ab0ceea90aaf7bea80199746fc43