[SDK] Improve transpose op errors and tests (#62157)
Improve the error messages and testing for the transpose op.

**Rank-zero tensors**

PyTorch treats rank-zero tensors as a no-op, effectively returning the
same shape, so long as the input dimensions are in the range [-1,0].
Let's follow that same behavior.

**Out-of-bounds dimensions**

Check for dimensions that are out of bounds for the input tensor shape,
and provide a useful error message which indicates the dimension which
caused the error, the expected bounds, and the value of the dimension
which caused the error. This error message is pretty similar to the
PyTorch error messages for the same case. PyTorch does not include the
name of the argument that caused the error, but that information is easy
to report, so let's report it.

Closes GEX-2259

MAX_GRAPH_API_ORIG_REV_ID: bb21d7fbade6f2fe5aea117c710cabbdf40aa4f7