[python-bindings] polish: Add `PythonObject` downcasting methods
This adds and improves the APIs for checked and unchecked downcasting of
`PythonObject`s that are expected to contain wrapped Mojo values.

This consolidates the existing downcasting logic in one place, on
`PythonObject`, instead of on the low level `PyObjectPtr` type.

When a Mojo type is bound for use from Python, an instance of that type
may be created within Python and then passed to a bound Mojo function
that accepts an arbitrary `PythonObject` argument. To access the raw
fields of the underlying Mojo type directly, there needs to be a way to
"downcast" that `PythonObject` into the Mojo value it's known to
contain.

This requires first checking that the runtime type of the `PythonObject`
is the bound Python type object that represents the original Mojo type,
which is done by the new safe `PythonObject.downcast_value_ptr()`
method.

MODULAR_ORIG_COMMIT_REV_ID: 727c31dc19f559eaecd5ebe58d4283f486b601e1