[Serving] Limit number of active request tasks
We find in our profiles that a large spike in incoming requests leads to
a delay in requests reaching our ModelWorker. It appears this is due to
too many tasks getting launched concurrently in the serving frontend. We
see improved performance if we limit the number of active tasks fielding
incoming requests with an asyncio.Sempaphore, but this is likely just a
band-aid.

MODULAR_ORIG_COMMIT_REV_ID: f06af18142fa469550d99df060fd3b5e7ff94a5d