[stdlib] Introduce `PyObjectFunction` wrapper to manage variadic binding signatures
To avoid having a large number of overloads in the binding builder
classes, this patch introduces a wrapper struct that implicitly converts
from a number of different functions signatures that can be registered
in the binding builders. As a side effect, a lot of similar logic is
deduplicated and brought into this new struct to be hidden from the rest
of the code. This struct could be generated, but for now it will be
managed manually; nonetheless, it has been implemented to be
particularly concise and easy to extend using AI assistants.

MODULAR_ORIG_COMMIT_REV_ID: c43659eeb118f3875214ea1df9896f27c93e46b6