[Serve] Do not prematurely terminate metrics client
The metric client is supposed to be around for most of the lifetime of
the process. When the 'with' ends, we terminate the metrics client, and
nothing after it will have its metrics recorded. Accordingly, it is
crucial to ensure that the bulk of the work we perform in a model worker
is inside this 'with'.

MODULAR_ORIG_COMMIT_REV_ID: b72f2ac6dcd4b3f6f5a1326d6f55bc23a013e6c8