[Pipelines] Fix Type Hinting in Vision Models for Deserialization
Previously, throughout the Vision models, we were using a
`Sequence[np.ndarray]` type hint. This is not deserializable, as it
cannot be deconstructed from data, as Sequence is not a concrete type.
This has been expanded to be a tuple, which caught a few inconsistencies
in which we were manually updating this value to a list in model code.

MODULAR_ORIG_COMMIT_REV_ID: 30b12a2f4c90bc95283351ddc9d317e6b689ef15