[KGEN] Rename MulNuw to MulNoWrap (#63469)
The MulNuw opcode was introduced to enable certain algebraic
simplifications in the graph compiler, with the assumption that the
operation is undefined on overflow. The opcode's name is misleading
since the graph compiler uses signed integers to represent dimensions,
and based on its name MulNuw should behave the same as Mul for signed
integers.

This change renames MulNuw to MulNoWrap, which is more inline with how
it is used in the graph compiler and how it is treated in the
ParamOperatorAttr cannonicalizations.

MAX_GRAPH_API_ORIG_REV_ID: 9281570ace51a6c38eb17baf90bf96aab2629dd4