[SDK] Remove insignificant trailing commas in calls (NFC)
Trailing commas are considered significant to Ruff and cause each
argument to appear on a different line. In a lot of cases this behavior
is not intended and the code is clearer when it appears on one line.
This PR removes trailing commas on calls when it is easy to
programmatically show they were not significant.

This was done with a libcst codemod not included in this PR.

MODULAR_ORIG_COMMIT_REV_ID: 0bd0538d155fd3279762f38d8548e90f62932c79