[bazel] Add support for building some open source python
This fixes some python macros to correctly thread these through how we
do internally as well. For parts of the SDK that are not open source we
pull those from the current version of the wheel.

There are still some stragglers that are excluded and definitely cause
runtime failures given the right inputs, but this is a start.

MODULAR_ORIG_COMMIT_REV_ID: 2282850f380f7cfa9458c843f4cd86b9bc75ca27