[mojo-stdlib] Introduce a new `IteratorTrait` trait
This patch renames the newly-sensible `_ParamForIterator` to
`IteratorTrait`, and renames its assoc alias to `Element`. This
moves the alias to value.mojo (next to some other traits) and
auto-exports it from the stdlib.

This is a small step forward for generic programming, and will allow
more types to work with `@parameter for` (which requires
conformance to this trait).  That said, there are a number of
questions that we need to decide on over time.  Here are a few
of them with loosely-held rationale for these design points, but
I defer to other stdlib experts to make the final calls:

1) Where does this live? I put it in value.mojo but we could have
    a new sequence.mojo or iterator.mojo.
2) Should there be a Sequence trait that things like List directly
    conform to, which requires an iter method returning something
    that conforms to `IteratorTrait`?
3) What should the actual name of this be?  I picked `IteratorTrait`
     following the swift `IteratorProtocol` name.  The name `Iterator`
     is one candidate, but there are problems, see below.
4) We also want "reference iteratable" things someday.  The origin
     system needs to make progress, but then we can have a new
     "RefiterableTrait" trait that refines IteratorTrait with the
     ref-returning next requirement.  With default impls in traits, we
     could even have it provide the "next" method.
5) This patch doesn't conform Span to IteratorTrait, because it is
     generic across address spaces and we can't do true conditional
     trait conformance yet.

On the naming thing, see the "Name variables, parameters, and
associated types according to their roles," part of the [Swift API
design
guidelines](https://www.swift.org/documentation/api-design-guidelines/#promote-clear-usage).
The basic problem is that you end up wanting thing like this, so
naming it "Iterator" would cause all sorts of mayhem:

```mojo
trait Sequence:
  alias Iterator : IteratorTrait
  ...
trait IteratorTrait: ...
```

MODULAR_ORIG_COMMIT_REV_ID: 07f96662beda03bbc3cd1d6c70ce3d5dd7367338