[Kernel] Update MatmulAllReduce to work with a dynamic number of partitions for the M axis.
In practice, the M dimension (seq-len) is dynamic, and we want to use more paritions as M increases.
But for N, we still want to keep static partitions.
Updates the kernel to take the partitions count as a static or dynamic value.
Also created a new kernel for MatmulAllReduce that try to infers automatically the best dimension / split size.

MODULAR_ORIG_COMMIT_REV_ID: bf3380661a7095b974c5777b72d7856312500010