[Kernels] Fix RNG seed management for token sampling
We need RNG during token sampling to perform weighted sampling on the
token probabilities output from the model. We initially just passed the
seed to our RNG engine but realized that we would generate the same
number each iteration if we don't change the seed/offset at each step.
Our current production code attempts to fix this by setting the offset
to the index of the highest probability token, but we realized that this
would lead to the same undesirable behavior if the same token has a high
probability multiple steps in a row.

This PR patched this to instead increment the seed by the generated
sequence length, so we're guaranteed to have a unique RNG value at each
sampling step.

MODULAR_ORIG_COMMIT_REV_ID: c418431d06c2ea78944210b61254a6e17eebf814