[Pipelines] Provide Tensor directly to compute_log_probabilities_new
np.ndarray matches the existing interface, but rules out the possibility
of using an existing on-device Tensor. Switching logits and
last_token_logits to Tensor allow eliminating a transfer.

The same argument could be made for other np.ndarrays in the parameter
list, but the logits tensors are the largest and their transfer
currently represents the biggest bottleneck, so it is most important to
address these. Future work could move other parameters to Tensor as well
but it is not as important for performance.

MODULAR_ORIG_COMMIT_REV_ID: da0fc8ba35fe926a16e33141f65f926ba23a2c28