[Pipelines] Switch Llama3 & DeepSeekV2 to new logprobs implementation
Llama3 and DeepSeekV2 are the only models that support logprobs at all,
so in effect this switches over all models from the old logprobs
implementation to the new one. In doing so we eliminate device transfers
for the logit tensors, which should provide a huge speedup.

Remaining clean-up from the transition to be dealt with in a follow-up
PR.

MODULAR_ORIG_COMMIT_REV_ID: 9c6e423434fb363edfc04cac29c06d5bfea3a97c