[SDK] Make target_num_new_tokens required / default
Make `target_num_new_tokens` a required argument with a default.
Previously this was doing the same thing but hiding the default in
pipelines/llm.py.

This is needed so that memory estimation can understand the maximum
number of tokens in a batch.
In the future, if we make this logic more complex we can move it to a
dedicated "target batch size" estimator.
But the logic can't live in the scheduler since memory estimation needs
it.

MODULAR_ORIG_COMMIT_REV_ID: 08bc29f22aa97944ee61dca6a780ff2d6a3b9fdf