[bazel] Enable C++ path mapping
This enables a mode in bazel that strips the unique
`bazel-out/k8-dbg-ST-53d8acb3742a` transition based output hashes to
instead be `k8-dbg` for all configs. This allows bazel to then
deduplicate actions that are otherwise identical besides the transition
they are under.

This will allow us to use a transition on python version to eventually
be able to build multiple nanobind library versions without having to
expand to N underlying targets. Which will also enable us to test
multiple python versions.

Without this change each supported python version would end up with
their own unique `bazel-out` directory name, which would then result in
us building the entire transitive dependency tree N times.

You can see this working because bazel prints the number of
`deduplicated` actions:

```
INFO: 11 processes: 21 action cache hit, 1 internal, 6 darwin-sandbox, 4 deduplicated.
```

Currently this is only enabled for compile actions, the notable missing
actions are:

- CppLink, not yet supported by bazel. This means that all binary links
will still be specific to the transition it's built under. This is
fine-ish since binaries are only inputs to tests, which will differ
anyways, or running, in which case you'll only build a single
configuration anyways.
- CppArchive, this one is supported but apparently causes issues with
debug info in generated source files that are part of the archive.
Having duplicated versions of these should be a no-op on linux because
we use `--start-lib/--end-lib` which means static archives aren't even
created in the build. On macOS this will result in duplicate archives
which isn't ideal, but since they're only inputs to CppLink, and that's
already not deduplicated, the primary downside would be downloading more
archives. If this becomes an issue we can try to enable it, because it's
possible we don't care too much about the debug info issues for
generated files

This breaks debug info for generated files unless we set a
target.source-map in lldb. For now I'm ignoring that but we can fix that
if someone hits it.

MODULAR_ORIG_COMMIT_REV_ID: d76ceca84e2e4a4b432f6c0c780102935e4a3cea