[Kernels] Add automatic casting in `fused_qk_ragged_rope`
This PR allows the `input` tensor is not of the same dtype as
`freqs_cis`. In this case, the `input` tensor will be cast to the dtype
of `freqs_cis` for the RoPE computation, and cast back to the original
dtype once the computation is finished.

Also, in Llama4 we no longer cast the `freqs_cis` to `bfloat16` but keep
it in `float32` to align with the reference Llama4 implementation
behavior.

MODULAR_ORIG_COMMIT_REV_ID: 547e93c876746df508d217e8fc2078a6cfd149b1