[bazel] Fix GPU tags validation
mojo_filecheck_test wasn't validating that it had the gpu tag when the
target was restricted to GPUs. With how we do this now we have to parse
the select() values to understand this since we can't just check if a
value is in that list. The logic for this has to use a runtime hack to
introspect those values, and if any of the potential values are GPU
related, assume it requires the GPU tag.

MODULAR_ORIG_COMMIT_REV_ID: 507dc112c383eb358c75b65238a64abfeaaf0d62