[stdlib] polish: Add `CompilationTarget.unsupported_target_error()` + fix `thread_idx.x` on CPU error
This is intended to establish a pattern for ensuring we emit descriptive
error messages when an `@parameter if ... else ...` chain does not
include a clause for the current target.

This updates the Mojo standard library style guide to document the
purpose of this pattern.

This fixes a specific instance of this pattern: when accessing
`thread_idx.x` when targeting a CPU, previously this error was emitted:

    LLVM ERROR: Cannot select: intrinsic %llvm.amdgcn.workitem.id.x
/bin/bash: line 1: 71243 Abort trap: 6 mojo build thread-idx-on-cpu.mojo

This error is bad because it doesn't include any useful information
about what went wrong, and it only happens because we (incorrectly) try
to emit an AMD GPU instruction when targeting a CPU.

After this fix, a elaboration constraint failure is emitted:

    stdlib/builtin/constrained.mojo:58:6:
note: constraint failed: Current compilation target does not support
operation: thread_idx field access
        ]()
     ^
MODULAR_ORIG_COMMIT_REV_ID: 2c2c2e13f791c09b793ae3d86c5e5eca59b3d8b7