[External] [stdlib] Remove `Bool.__neg__` (#65738)
[External] [stdlib] Remove `Bool.__neg__`

Mojo is a language with strong typing, there's no point mimic Python
oddity here.

> I'd suggest we remove that behavior since Mojo has strong typing. I
remember me and you were talking about this last year sometime and it
made one particular thing easier, but it is not a great thing for Mojo
to keep from Python IMO. We've moved on from the "superset of Python"
and needing to copy some of its not-so-great things. This is one of
those things we should deviate from IMO. Any interest in opening a PR?

_Originally posted by @JoeLoser in
https://github.com/modular/modular/pull/4898#discussion_r2168080582_

Co-authored-by: soraros <soraros@users.noreply.github.com>
Closes modular/modular#5044
MODULAR_ORIG_COMMIT_REV_ID: eb3c7e547833111dbafd4acb3c8b01dbcd6d0e82