[stdlib] Rename `DeviceContext.device_info` to `default_device_info`
This name was extremely misleading: when users would access the
`device_info` field on a `DeviceContext` object, they might think they
are getting the info associated with that context, when in fact they
were getting the default info for the accelerator target. This happened
to be mostly okay, since `DeviceContext` instances almost always refer
to the current target accelerator`, but this will become a problem if
more than one accelerator type is present, and the field is ambiguous if
the target API is `"cpu"`. This change also opens up the opportunity to
change `DeviceContext` to explicitly model parametric device info.

MODULAR_ORIG_COMMIT_REV_ID: 74c49cd1f9043f3d8208e1e4fc01ac0fc071e611